Move existing behavior into StandaloneAction
This clears the way for us to add an ApplicationAction (providing Hanami application integration behavior) in front of StandaloneAction in the ancestor tree.
